# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Country Decision Atlas
  version: 0.1.0
paths:
...
  /api/v1/personas:
    get:
      tags:
      - personas
      summary: Read Personas
      operationId: read_personas_api_v1_personas_get
      parameters:
      - name: locale
        in: query
        required: false
        schema:
          type: string
          pattern: ^[a-z]{2}$
          enum:
          - en
          - ru
          default: ru
          title: Locale
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersonaListResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/personas/{persona_slug}:
    get:
      tags:
      - personas
      summary: Read Persona
      operationId: read_persona_api_v1_personas__persona_slug__get
      parameters:
      - name: persona_slug
        in: path
        required: true
        schema:
          type: string
          title: Persona Slug
      - name: locale
        in: query
        required: false
        schema:
          type: string
          pattern: ^[a-z]{2}$
          enum:
          - en
          - ru
          default: ru
          title: Locale
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Persona'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/personas/{persona_slug}/weights:
    get:
      tags:
      - personas
      summary: Read Persona Weights
      operationId: read_persona_weights_api_v1_personas__persona_slug__weights_get
      parameters:
      - name: persona_slug
        in: path
        required: true
        schema:
          type: string
          title: Persona Slug
      - name: scenario
        in: query
        required: true
        schema:
          type: string
          description: Scenario slug
          title: Scenario
        description: Scenario slug
      - name: version
        in: query
        required: false
        schema:
          type: string
          pattern: ^v\d+\.\d+$
          default: v1.0
          title: Version
      - name: locale
        in: query
        required: false
        schema:
          type: string
          pattern: ^[a-z]{2}$
          enum:
          - en
          - ru
          default: ru
          title: Locale
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersonaWeightProfileResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    LocaleCode:
      type: string
      enum:
      - en
      - ru
      title: LocaleCode
    LocaleResolution:
      properties:
        requested_locale:
          $ref: '#/components/schemas/LocaleCode'
        resolved_locale:
          $ref: '#/components/schemas/LocaleCode'
        translation_status:
          $ref: '#/components/schemas/TranslationStatus'
      type: object
      required:
      - requested_locale
      - resolved_locale
      - translation_status
      title: LocaleResolution
...
    Persona:
      properties:
        slug:
          type: string
          title: Slug
        name:
          type: string
          title: Name
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        is_active:
          type: boolean
          title: Is Active
          default: true
        display_order:
          type: integer
          title: Display Order
      type: object
      required:
      - slug
      - name
      - display_order
      title: Persona
    PersonaAdjustedMetricWeight:
      properties:
        metric_id:
          type: string
          title: Metric Id
        metric_slug:
          type: string
          title: Metric Slug
        metric_name:
          type: string
          title: Metric Name
        base_weight:
          type: number
          title: Base Weight
        modifier:
          type: number
          title: Modifier
        adjusted_weight:
          type: number
          title: Adjusted Weight
      type: object
      required:
      - metric_id
      - metric_slug
      - metric_name
      - base_weight
      - modifier
      - adjusted_weight
      title: PersonaAdjustedMetricWeight
    PersonaListResponse:
      properties:
        items:
          items:
            $ref: '#/components/schemas/Persona'
          type: array
          title: Items
        locale:
          $ref: '#/components/schemas/LocaleResolution'
      type: object
      required:
      - locale
      title: PersonaListResponse
    PersonaWeightProfile:
      properties:
        persona:
          $ref: '#/components/schemas/Persona'
        scenario_slug:
          type: string
          title: Scenario Slug
        version:
          type: string
          title: Version
        weights:
          items:
            $ref: '#/components/schemas/PersonaAdjustedMetricWeight'
          type: array
          title: Weights
        weight_sum:
          type: number
          title: Weight Sum
      type: object
      required:
      - persona
      - scenario_slug
      - version
      - weight_sum
      title: PersonaWeightProfile
    PersonaWeightProfileResponse:
      properties:
        item:
          $ref: '#/components/schemas/PersonaWeightProfile'
      type: object
      required:
      - item
      title: PersonaWeightProfileResponse
...
    TranslationStatus:
      type: string
      enum:
      - source
      - translated
      - fallback
      - missing
      title: TranslationStatus
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    AdminTokenAuth:
      type: apiKey
      in: header
      name: X-Admin-Token